move profile.pb.go , profile.proto to pkg tree
